src/main/java/com/projet/controller/CustomerController.java
package com.projet.controller;

import com.projet.entities.Client;
import com.projet.metier.CustomerMetierInterface;

import lombok.extern.slf4j.Slf4j;
import net.sf.jasperreports.engine.*;
import net.sf.jasperreports.engine.data.JRBeanCollectionDataSource;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.util.ResourceUtils;
import org.springframework.web.bind.annotation.*;

import javax.management.JMRuntimeException;
import java.io.File;
import java.io.FileNotFoundException;
import java.util.*;

@RestController
@RequestMapping("/clients")
@Slf4j
@CrossOrigin(origins = "http://localhost:4200")
public class CustomerController {


    CustomerMetierInterface customerMetier;

    @Autowired
    public CustomerController(CustomerMetierInterface customerMetier) {
        this.customerMetier = customerMetier;
    }

    @GetMapping(value = "/index")
    public String accueil() {
        return "BienVenue au service Web REST 'produits'.....";
    }

    @PostMapping(
            value = "/add", produces = {MediaType.APPLICATION_JSON_VALUE,
            MediaType.APPLICATION_XML_VALUE}
    )
    public Client saveProduit(@RequestBody Client p) {
        // Check if the image data is present

        return customerMetier.save(p);
    }

    @GetMapping(value = "/", produces = MediaType.APPLICATION_JSON_VALUE)
    public List<Client> getAllProduits() {

        return customerMetier.listeClients();
    }

    @GetMapping(value = "/Actif", produces = MediaType.APPLICATION_JSON_VALUE)
    public List<Client> getActifCustomers() {

        return customerMetier.getActifCustomers();
    }

    @GetMapping(value = "/NonActif", produces = MediaType.APPLICATION_JSON_VALUE)
    public List<Client> getNonActif() {
        return customerMetier.getNonActifCustomers();
    }



    @PutMapping(value = "/update/{customerId}", produces = MediaType.APPLICATION_JSON_VALUE)
    public Client updateCustomer(@PathVariable Long customerId, @RequestBody Client clt) {
        clt.setIdCl(customerId);
        return customerMetier.updateClient(clt);
    }


    @DeleteMapping(value = "/delete/{customerId}", produces = MediaType.APPLICATION_JSON_VALUE)
    public void deleteProduit(@PathVariable Long customerId) {
        customerMetier.deleteCustomer(customerId);
    }

    @GetMapping(value = "{customerId}", produces = MediaType.APPLICATION_JSON_VALUE)
    public Optional<Client> getCltId(@PathVariable Long customerId) {
        return customerMetier.findCltId(customerId);
    }

    @GetMapping(value = "/search", produces = MediaType.APPLICATION_JSON_VALUE)
    public List<Client> searchCustomers(@RequestParam(name = "keyword") String keyword) {

        return customerMetier.searchCustomers("%" + keyword + "%");
    }

    @GetMapping(value = "/searchadd", produces = MediaType.APPLICATION_JSON_VALUE)
    public List<Client> searchCustomersByAdd(@RequestParam(name = "adresse") String adresse) {

        return customerMetier.searchCustomerByadd("%" + adresse + "%");
    }

    @GetMapping(value = "/searchtel", produces = MediaType.APPLICATION_JSON_VALUE)
    public List<Client> searchCustomersByNumTel(@RequestParam(name = "NumTel") String NumTel) {

        return customerMetier.searchCustomerByNumTel("%" + NumTel + "%");
    }
    @GetMapping(value = "/{clientId}/getTotalRevenue", produces = MediaType.APPLICATION_JSON_VALUE)
    public Double getTotalRevenueForClient(@PathVariable Long clientId) {
        Optional<Client> client = customerMetier.findCltId(clientId);
        if (client.isPresent())
            return customerMetier.getTotalRevenueForClient(client.get());

        return 0.0;
    }


    @GetMapping(value = "/getTotalRevenue", produces = MediaType.APPLICATION_JSON_VALUE)
    public List<Object[]> getTotalRevenueForAllClient() {
        return customerMetier.getTotalRevenueForAllClients();


    }

    @GetMapping(value = "/getRemainingAmount", produces = MediaType.APPLICATION_JSON_VALUE)
    public List<Object[]> getRemainingAmountAllClient() {
        return customerMetier.getRemainingAmountForAllClients();
    }

    @GetMapping(value = "/{clientId}/getRemainingAmount", produces = MediaType.APPLICATION_JSON_VALUE)
    public Double getRemainingAmountForClient(@PathVariable Long clientId) {
        Optional<Client> client = customerMetier.findCltId(clientId);
        if (client.isPresent())
            return customerMetier.getRemainingAmountForClient(client.get());

        return 0.0;
    }

    @GetMapping("/paymentsClients")
    public List<Object[]> getPaymentsStatusForAllClients() {
        return customerMetier.getPaymentsStatusForAllClients();
    }

    @GetMapping("/{clientId}/revenue/{year}")
    public List<Object[]> getRevenueByClientAndYear(@PathVariable Long clientId, @PathVariable String year) {
        // Retrieve the client by ID
        Client client = customerMetier.findCltId(clientId).get();

        return customerMetier.getRevenueByClientAndYear(client, year);
    }

    @GetMapping("/revenue/{year}")
    public List<Object[]> getRevenueByYear(@PathVariable String year) {
        return customerMetier.getRevenueByYear(year);
    }

    @GetMapping("/imprimer")
    public ResponseEntity<byte[]> downloadListCustomersVal() throws FileNotFoundException, JRException {
        File file = ResourceUtils.getFile("classpath:listeClients.jrxml");
        JasperReport jasperReport = JasperCompileManager.compileReport(file.getAbsolutePath());
        List<Client> liste = customerMetier.listeClients();

        System.out.println(liste.size());

        JRBeanCollectionDataSource dataSource = new JRBeanCollectionDataSource(liste);

        JasperPrint jasperPrint = JasperFillManager.fillReport(jasperReport, null, dataSource);

        return ResponseEntity
                .ok()
                .header("Content-Type", "application/pdf; charset=UTF-8")
                .header("Content-Disposition", "inline; filename=\"" + "downloadListCustomersList" + ".pdf\"")
                .body(JasperExportManager.exportReportToPdf(jasperPrint));
    }



}
